requirements.txt
pandas==2.1.4
selenium==3.141.0
python-dotenv==1.0.1